fix grTexDownloadMipMapLevelPartial - it should only read and replace data of the size specified by start-end rows
Previously the start-end range was ignored and data block of the size of the given mip level was copied to internal memory.
This works only if start is 0, the data block is large enough and the rest of the texture data is unchanged.
Otherwise it causes texture corruption or crashes due to invalid reads.