Remove the MMX version of memcpy()
The libc version is most likely a lot better optimised towards modern
CPUs which support newer SIMD extensions.

On my Kaby Lake CPU, in a simple benchmark, this MMX version would reach
25 GiB/s for size=4096 and dip to 13 GiB/s at higher sizes, while
glibc’s memcpy() would reach 90 GiB/s and go down to 31 GiB/s minimum.

Of course this is a very unscientific benchmark, not accounting for
caches and such, but it gives an estimation of the gains to be had when
removing this function.