Avoid including multiple different glext.h versions
platform/window.h includes GL/gl.h which in turns includes GL/glext.h.
This could be fine and all, if its include guard was the same as our
glext.h, alas it is __gl_glext_h__ instead of __glext_h__ so both get
included.

This isn’t much of an issue since they are supposed to provide the same
extensions, but redefining GL_GLEXT_VERSION still triggers a warning.

In the future it would be nice to revisit whether or not it’s useful to
ship our own glext.h when the system already provides one.